@@ -4,6 +4,7 @@ namespace App\Http\Middleware;
 
 use Closure;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class StaffMiddleware
@@ -15,9 +16,9 @@ class StaffMiddleware
      * @param  \Closure  $next
      * @return mixed
      */
-    public function handle(Request $request, Closure $next)
+    public function handle(Request $request, Closure $next): Response
     {
-        if (!auth()->check() || !auth()->user()->isStaff()) {
+        if (!Auth::check() || !Auth::user()->isStaff()) {
             abort(403, 'Bu sayfaya erişim yetkiniz bulunmamaktadır.');
         }
 
