@@ -10,6 +10,11 @@ class Borrowing extends Model
 {
     use HasFactory;
 
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array<int, string>
+     */
     protected $fillable = [
         'user_id',
         'book_id',
@@ -18,10 +23,22 @@ class Borrowing extends Model
         'due_date',
         'returned_at',
         'condition',
-        'status',
         'notes',
-        'reject_reason',
-        'fine_amount'
+        'fine_amount',
+        'status'
+    ];
+
+    protected $dates = [
+        'borrow_date',
+        'due_date',
+        'returned_at',
+        'created_at',
+        'updated_at'
+    ];
+
+    protected $attributes = [
+        'status' => 'pending',
+        'fine_amount' => 0.00
     ];
 
     protected $casts = [
@@ -30,6 +47,29 @@ class Borrowing extends Model
         'returned_at' => 'datetime',
         'fine_amount' => 'decimal:2'
     ];
+
+    // Borrowing statuses
+    const STATUS_PENDING = 'pending';
+    const STATUS_ACTIVE = 'active';
+    const STATUS_RETURNED = 'returned';
+    const STATUS_OVERDUE = 'overdue';
+    const STATUS_LOST = 'lost';
+
+    // Book conditions
+    const CONDITION_GOOD = 'good';
+    const CONDITION_DAMAGED = 'damaged';
+    const CONDITION_LOST = 'lost';
+
+    protected static function boot()
+    {
+        parent::boot();
+
+        static::creating(function ($borrowing) {
+            if (!isset($borrowing->fine_amount)) {
+                $borrowing->fine_amount = 0.00;
+            }
+        });
+    }
 
     public function user()
     {
@@ -46,42 +86,86 @@ class Borrowing extends Model
         return $this->belongsTo(Stock::class);
     }
 
+    public function approver()
+    {
+        return $this->belongsTo(User::class, 'approved_by');
+    }
+
+    public function rejecter()
+    {
+        return $this->belongsTo(User::class, 'rejected_by');
+    }
+
+    public function returner()
+    {
+        return $this->belongsTo(User::class, 'returned_to');
+    }
+
+    public function fines()
+    {
+        return $this->hasMany(Fine::class);
+    }
+
     public function isOverdue()
     {
         if ($this->returned_at) {
-            return $this->returned_at->isAfter($this->due_date);
+            return $this->returned_at > $this->due_date;
         }
-        
-        return now()->isAfter($this->due_date);
+        return now() > $this->due_date;
     }
-    
+
+    public function getDaysOverdueAttribute()
+    {
+        if (!$this->isOverdue()) {
+            return 0;
+        }
+
+        return now()->diffInDays($this->due_date);
+    }
+
     public function getStatus()
     {
-        if ($this->isOverdue()) {
-            return 'overdue';
+        if ($this->returned_at) {
+            return self::STATUS_RETURNED;
         }
-        
+
+        if ($this->isOverdue()) {
+            return self::STATUS_OVERDUE;
+        }
+
         return $this->status;
     }
-    
+
+    public function getStatusLabelAttribute()
+    {
+        return match($this->getStatus()) {
+            self::STATUS_PENDING => 'Beklemede',
+            self::STATUS_ACTIVE => 'Aktif',
+            self::STATUS_RETURNED => 'İade Edildi',
+            self::STATUS_OVERDUE => 'Gecikmiş',
+            self::STATUS_LOST => 'Kayıp',
+            default => 'Bilinmiyor'
+        };
+    }
+
     public function isPending()
     {
-        return $this->status === 'pending';
+        return $this->status === self::STATUS_PENDING;
     }
-    
-    public function isApproved()
+
+    public function isActive()
     {
-        return $this->status === 'approved';
+        return $this->status === self::STATUS_ACTIVE;
     }
-    
-    public function isRejected()
-    {
-        return $this->status === 'rejected';
-    }
-    
+
     public function isReturned()
     {
-        return $this->status === 'returned' || !is_null($this->returned_at);
+        return $this->status === self::STATUS_RETURNED;
+    }
+
+    public function isLost()
+    {
+        return $this->status === self::STATUS_LOST;
     }
 
     /**
@@ -94,7 +178,7 @@ class Borrowing extends Model
         }
 
         $endDate = $this->returned_at ?? now();
-        return $this->due_date->diffInDays($endDate, false);
+        return $endDate->diffInDays($this->due_date);
     }
 
     /**
@@ -105,17 +189,49 @@ class Borrowing extends Model
         return $this->getOverdueDays() * $dailyRate;
     }
 
+    public function updateFineAmount()
+    {
+        $this->fine_amount = $this->calculateFine();
+        return $this->save();
+    }
+
     /**
      * Kitabı iade et
      */
-    public function returnBook($condition = 'good')
+    public function returnBook($condition = self::CONDITION_GOOD)
     {
         $this->returned_at = now();
         $this->condition = $condition;
-        $this->status = 'returned';
+        $this->status = self::STATUS_RETURNED;
         
+        // Stok durumunu güncelle
+        if ($this->stock) {
+            $this->stock->status = match($condition) {
+                self::CONDITION_DAMAGED => Stock::STATUS_DAMAGED,
+                self::CONDITION_LOST => Stock::STATUS_LOST,
+                default => Stock::STATUS_AVAILABLE
+            };
+            
+            $this->stock->is_available = $condition === self::CONDITION_GOOD;
+            $this->stock->save();
+        }
+
+        // Gecikme cezası kontrolü
         if ($this->isOverdue()) {
-            $this->fine_amount = $this->calculateFine();
+            $overdueDays = $this->getOverdueDays();
+            $fineAmount = $overdueDays * (session('overdue_fine_per_day') ?? 1.0);
+
+            $fine = new Fine([
+                'user_id' => $this->user_id,
+                'book_id' => $this->book_id,
+                'amount' => $fineAmount,
+                'reason' => 'late',
+                'days_late' => $overdueDays,
+                'status' => Fine::STATUS_PENDING
+            ]);
+
+            $this->fines()->save($fine);
+            $this->fine_amount = $fineAmount;
         }
 
         return $this->save();
