@@ -6,17 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    public function up()
+    public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->boolean('is_staff')->default(false)->after('is_admin');
-        });
+        if (!Schema::hasColumn('users', 'is_staff')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->boolean('is_staff')->default(false)->after('is_admin');
+            });
+        }
     }
 
-    public function down()
+    public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('is_staff');
-        });
+        if (Schema::hasColumn('users', 'is_staff')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('is_staff');
+            });
+        }
     }
 }; 
