@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('stocks', function (Blueprint $table) {
-            $table->decimal('acquisition_price', 10, 2)->nullable();
-        });
+        if (!Schema::hasColumn('stocks', 'acquisition_price')) {
+            Schema::table('stocks', function (Blueprint $table) {
+                $table->decimal('acquisition_price', 10, 2)->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('stocks', function (Blueprint $table) {
-            $table->dropColumn('acquisition_price');
-        });
+        if (Schema::hasColumn('stocks', 'acquisition_price')) {
+            Schema::table('stocks', function (Blueprint $table) {
+                $table->dropColumn('acquisition_price');
+            });
+        }
     }
 };
