@@ -3,48 +3,39 @@
 @section('title', 'Yeni Ödünç Verme')
 
 @section('styles')
-<link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css" rel="stylesheet" />
-<link href="https://cdn.jsdelivr.net/npm/select2-bootstrap4-theme@1.0.0/dist/select2-bootstrap4.min.css" rel="stylesheet" />
 <style>
-    .select2-container {
-        z-index: 9999;
-    }
-    .select2-dropdown {
-        z-index: 10000;
-    }
-    .select2-results {
-        max-height: 300px;
-        overflow-y: auto;
-    }
-    /* Normal select için stil */
-    select.form-control {
-        width: 100%;
-        padding: 8px;
-        border-radius: 4px;
-        border: 1px solid #ced4da;
-        height: auto;
-    }
     .search-results {
-        max-height: 200px;
+        max-height: 300px;
         overflow-y: auto;
         border: 1px solid #ddd;
         border-radius: 4px;
-        display: none;
+        background: white;
+        z-index: 1000;
+        position: absolute;
+        width: 100%;
     }
     .search-item {
-        padding: 10px;
-        cursor: pointer;
+        padding: 10px 15px;
         border-bottom: 1px solid #eee;
+        cursor: pointer;
+        transition: background-color 0.2s;
     }
     .search-item:hover {
         background-color: #f8f9fa;
+    }
+    .book-title {
+        font-weight: bold;
+        margin-bottom: 5px;
+    }
+    .book-info {
+        font-size: 0.9em;
+        color: #666;
     }
     .selected-item {
         padding: 15px;
         border: 1px solid #ddd;
         border-radius: 4px;
         margin-top: 10px;
-        display: none;
     }
 </style>
 @endsection
@@ -63,111 +54,111 @@
             <h6 class="m-0 font-weight-bold text-primary">Ödünç Bilgileri</h6>
         </div>
         <div class="card-body">
-            @if($errors->any())
-                <div class="alert alert-danger">
-                    <ul class="mb-0">
-                        @foreach($errors->all() as $error)
-                            <li>{{ $error }}</li>
-                        @endforeach
-                    </ul>
-                </div>
-            @endif
-
-            <form id="borrowingForm" action="{{ route('staff.borrowings.store') }}" method="POST">
+            <form id="borrowingForm" method="POST" action="{{ route('staff.borrowings.store') }}">
                 @csrf
-                
-                <!-- Kullanıcı Arama -->
-                <div class="mb-4">
-                    <label class="form-label">Kullanıcı Ara <span class="text-danger">*</span></label>
-                    <div class="input-group">
-                        <input type="text" class="form-control" id="userSearch" placeholder="İsim veya email ile kullanıcı ara...">
-                        <button class="btn btn-outline-secondary d-none" type="button" id="clearUserSearch">
-                            <i class="fas fa-times"></i>
-                        </button>
+                <div class="row">
+                    <!-- Kullanıcı Seçimi -->
+                    <div class="col-md-6 mb-3">
+                        <label class="form-label">Kullanıcı <span class="text-danger">*</span></label>
+                        <div class="user-search-container position-relative">
+                            <div class="input-group">
+                                <input type="text" class="form-control" id="userSearch" placeholder="Ad, soyad, e-posta veya telefon ile ara...">
+                                <button class="btn btn-outline-secondary" type="button" id="searchUserBtn">
+                                    <i class="fas fa-search"></i>
+                                </button>
                             </div>
-                    <div id="userSearchResults" class="search-results"></div>
-                    <div id="selectedUser" class="selected-item">
-                        <input type="hidden" name="user_id" id="selectedUserId">
-                        <div class="d-flex justify-content-between align-items-center">
-                            <div>
-                                <h6 class="mb-1" id="selectedUserName"></h6>
-                                <small class="text-muted" id="selectedUserEmail"></small>
+                            <input type="hidden" name="user_id" id="selectedUserId">
+                            <div id="userSearchResults" class="search-results" style="display: none;"></div>
+                            <div id="selectedUserDetails" class="selected-item" style="display: none;">
+                                <div class="d-flex justify-content-between align-items-start">
+                                    <div>
+                                        <h6 class="mb-1" id="selectedUserName"></h6>
+                                        <p class="mb-1" id="selectedUserEmail"></p>
+                                        <p class="mb-1" id="selectedUserPhone"></p>
+                                        <p class="mb-0" id="selectedUserMemberSince"></p>
+                                        <div id="selectedUserCurrentBooks" class="mt-2"></div>
+                                    </div>
+                                    <button type="button" class="btn btn-sm btn-outline-danger" id="clearUserSelection">
+                                        <i class="fas fa-times"></i>
+                                    </button>
+                                </div>
                             </div>
-                            <button type="button" class="btn btn-sm btn-outline-danger" id="clearSelectedUser">
-                                <i class="fas fa-times"></i>
-                            </button>
+                        </div>
                     </div>
-                    </div>
-                </div>
 
-                <!-- Kitap Arama -->
-                <div class="mb-4">
-                    <label class="form-label">ISBN ile Kitap Ara <span class="text-danger">*</span></label>
-                    <div class="input-group">
-                        <input type="text" class="form-control" id="isbnSearch" placeholder="ISBN numarası girin...">
-                        <button class="btn btn-outline-secondary" type="button" id="searchBookBtn">
-                            <i class="fas fa-search"></i>
-                        </button>
-                        <button class="btn btn-outline-secondary d-none" type="button" id="clearBookSearch">
-                            <i class="fas fa-times"></i>
-                        </button>
-                    </div>
-                    <div id="bookSearchFeedback" class="mt-2" style="display: none;">
-                        <div class="alert alert-danger"></div>
-                    </div>
-                    <div id="selectedBook" class="selected-item">
-                        <input type="hidden" name="book_id" id="selectedBookId">
-                        <div class="d-flex justify-content-between align-items-center">
-                            <div>
-                                <h6 class="mb-1" id="selectedBookTitle"></h6>
-                                <small class="text-muted">ISBN: <span id="selectedBookIsbn"></span></small><br>
-                                <small class="text-muted">Yazar(lar): <span id="selectedBookAuthor"></span></small><br>
-                                <small class="text-muted">Yayınevi: <span id="selectedBookPublisher"></span></small><br>
-                                <small class="text-muted">Kategori: <span id="selectedBookCategory"></span></small><br>
-                                <small class="text-muted">Sayfa Sayısı: <span id="selectedBookPages"></span></small><br>
-                                <small class="text-muted">Basım Yılı: <span id="selectedBookYear"></span></small><br>
-                                <small class="text-muted">Uygun Kopya Sayısı: <span id="selectedBookAvailableStock"></span></small>
+                    <!-- Kitap Seçimi -->
+                    <div class="col-md-6 mb-3">
+                        <label class="form-label">Kitap Ara <span class="text-danger">*</span></label>
+                        <div class="book-search-container position-relative">
+                            <div class="input-group">
+                                <input type="text" class="form-control" id="bookSearch" placeholder="ISBN veya kitap adı ile ara...">
+                                <button class="btn btn-outline-secondary" type="button" id="searchBookBtn">
+                                    <i class="fas fa-search"></i>
+                                </button>
                             </div>
-                            <button type="button" class="btn btn-sm btn-outline-danger" id="clearSelectedBook">
-                                <i class="fas fa-times"></i>
-                            </button>
+                            <input type="hidden" name="book_id" id="selectedBookId">
+                            <div id="bookSearchResults" class="search-results" style="display: none;"></div>
+                            <div id="selectedBookDetails" class="selected-item" style="display: none;">
+                                <div class="d-flex justify-content-between align-items-start">
+                                    <div>
+                                        <h6 class="mb-1" id="selectedBookTitle"></h6>
+                                        <p class="mb-1" id="selectedBookAuthor"></p>
+                                        <p class="mb-1" id="selectedBookPublisher"></p>
+                                        <p class="mb-0" id="selectedBookIsbn"></p>
+                                    </div>
+                                    <button type="button" class="btn btn-sm btn-outline-danger" id="clearBookSelection">
+                                        <i class="fas fa-times"></i>
+                                    </button>
+                                </div>
+                                <div class="mt-2">
+                                    <label class="form-label">Kopya Seçimi:</label>
+                                    <select class="form-select" id="stockSelect" name="stock_id" required>
+                                        <option value="">Kopya seçiniz...</option>
+                                    </select>
+                                </div>
+                            </div>
                         </div>
                     </div>
                 </div>
-                
-                <!-- Ödünç Alma Tarihi -->
-                <div class="mb-4">
-                    <label for="borrow_date" class="form-label">Ödünç Alma Tarihi <span class="text-danger">*</span></label>
-                    <input type="date" class="form-control" id="borrow_date" name="borrow_date" required 
-                           value="{{ date('Y-m-d') }}">
-                </div>
 
-                <!-- Teslim Tarihi -->
-                <div class="mb-4">
-                    <label for="due_date" class="form-label">Son Teslim Tarihi <span class="text-danger">*</span></label>
-                    <input type="date" class="form-control" id="due_date" name="due_date" required 
-                           min="{{ date('Y-m-d', strtotime('+1 day')) }}">
+                <div class="row">
+                    <!-- Ödünç Tarihi -->
+                    <div class="col-md-6 mb-3">
+                        <label class="form-label">Ödünç Tarihi <span class="text-danger">*</span></label>
+                        <input type="date" class="form-control" name="borrow_date" value="{{ date('Y-m-d') }}" required>
+                    </div>
+
+                    <!-- İade Tarihi -->
+                    <div class="col-md-6 mb-3">
+                        <label class="form-label">Son İade Tarihi <span class="text-danger">*</span></label>
+                        <input type="date" class="form-control" name="due_date" value="{{ date('Y-m-d', strtotime('+15 days')) }}" required>
+                    </div>
                 </div>
 
                 <!-- Notlar -->
-                <div class="mb-4">
-                    <label for="notes" class="form-label">Notlar</label>
-                    <textarea class="form-control" id="notes" name="notes" rows="3"></textarea>
+                <div class="mb-3">
+                    <label class="form-label">Notlar</label>
+                    <textarea class="form-control" name="notes" rows="3"></textarea>
                 </div>
 
                 <!-- Otomatik Onay -->
-                <div class="mb-4">
+                <div class="mb-3">
                     <div class="form-check">
-                    <input class="form-check-input" type="checkbox" id="auto_approve" name="auto_approve" value="1" checked>
-                    <label class="form-check-label" for="auto_approve">
-                        Otomatik Onayla
-                    </label>
-                </div>
+                        <input class="form-check-input" type="checkbox" name="auto_approve" id="autoApprove" value="1" checked>
+                        <label class="form-check-label" for="autoApprove">
+                            Otomatik Onayla
+                        </label>
+                    </div>
                 </div>
 
-                <button type="submit" class="btn btn-primary" id="submitBtn" disabled>
-                    <i class="fas fa-save"></i> Kaydet
-                </button>
+                <div class="d-flex justify-content-between">
+                    <a href="{{ route('staff.borrowings.index') }}" class="btn btn-secondary">
+                        <i class="fas fa-arrow-left"></i> Geri Dön
+                    </a>
+                    <button type="submit" class="btn btn-primary" id="submitBtn" disabled>
+                        <i class="fas fa-save"></i> Kaydet
+                    </button>
+                </div>
             </form>
         </div>
     </div>
@@ -175,211 +166,265 @@
 @endsection
 
 @section('scripts')
-<script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/js/select2.min.js"></script>
 <script>
 $(document).ready(function() {
     // Kullanıcı arama
     let userSearchTimeout;
-    $('#userSearch').on('input', function() {
+    
+    $('#userSearch').on('keyup', function() {
         clearTimeout(userSearchTimeout);
         const searchTerm = $(this).val().trim();
         
-        // Temizle butonunu göster/gizle
-        if (searchTerm.length > 0) {
-            $('#clearUserSearch').removeClass('d-none');
-        } else {
-            $('#clearUserSearch').addClass('d-none');
-        }
-        
         if (searchTerm.length < 2) {
-            $('#userSearchResults').empty().hide();
+            $('#userSearchResults').hide();
             return;
         }
-
+        
         userSearchTimeout = setTimeout(function() {
-            $.ajax({
-                url: '/staff/borrowings/search-user',
-                method: 'GET',
-                data: { term: searchTerm },
-                beforeSend: function() {
-                    $('#userSearchResults').html('<div class="p-2 text-center"><i class="fas fa-spinner fa-spin"></i> Aranıyor...</div>').show();
-                },
-                success: function(data) {
-                    if (data.length === 0) {
-                        $('#userSearchResults').html('<div class="p-2 text-center text-muted">Kullanıcı bulunamadı</div>');
+            $.get("{{ route('staff.borrowings.search.user') }}", { search: searchTerm })
+                .done(function(users) {
+                    if (users.length === 0) {
+                        $('#userSearchResults').html('<div class="p-3 text-center">Kullanıcı bulunamadı</div>').show();
                         return;
                     }
-
-                    let html = '';
-                    data.forEach(user => {
+                    
+                    let html = '<div class="list-group">';
+                    users.forEach(function(user) {
+                        let statusClass = user.can_borrow ? 'text-success' : 'text-danger';
+                        let statusText = '';
+                        
+                        if (!user.can_borrow) {
+                            if (user.overdue_books > 0) {
+                                statusText = `${user.overdue_books} gecikmiş kitap`;
+                            } else {
+                                statusText = `Maksimum kitap limitine ulaşıldı (${user.active_books}/4)`;
+                            }
+                        } else {
+                            statusText = `${user.active_books}/4 kitap`;
+                        }
+                            
                         html += `
-                            <div class="search-item" data-id="${user.id}" data-name="${user.name}" data-email="${user.email}">
-                                <strong>${user.name}</strong><br>
-                                <small class="text-muted">${user.email}</small>
-                            </div>
-                        `;
+                            <a href="#" class="list-group-item list-group-item-action user-result ${!user.can_borrow ? 'disabled' : ''}"
+                               data-user='${JSON.stringify(user)}'>
+                                <div class="d-flex justify-content-between align-items-start">
+                                    <div>
+                                        <h6 class="mb-1">${user.name}</h6>
+                                        <small class="text-muted d-block">
+                                            ${user.email}<br>
+                                            ${user.phone ? 'Tel: ' + user.phone : ''}<br>
+                                            Üyelik: ${user.member_since}
+                                        </small>
+                                    </div>
+                                    <small class="${statusClass}">${statusText}</small>
+                                </div>
+                            </a>`;
                     });
-                    $('#userSearchResults').html(html);
-                },
-                error: function(xhr) {
-                    $('#userSearchResults').html('<div class="p-2 text-center text-danger">Arama sırasında bir hata oluştu</div>');
-                    console.error('Kullanıcı arama hatası:', xhr.responseText);
-                }
-            });
-        }, 300);
-    });
-
-    // ISBN arama alanı için input olayı
-    $('#isbnSearch').on('input', function() {
-        const searchTerm = $(this).val().trim();
-        
-        // Temizle butonunu göster/gizle
-        if (searchTerm.length > 0) {
-            $('#clearBookSearch').removeClass('d-none');
-        } else {
-            $('#clearBookSearch').addClass('d-none');
-        }
+                    html += '</div>';
+                    
+                    $('#userSearchResults').html(html).show();
+                })
+                .fail(function(error) {
+                    console.error('Kullanıcı arama hatası:', error);
+                    $('#userSearchResults').html('<div class="p-3 text-center text-danger">Arama sırasında bir hata oluştu</div>').show();
+                });
+        }, 500);
     });
 
     // Kullanıcı seçimi
-    $(document).on('click', '#userSearchResults .search-item', function() {
-        const id = $(this).data('id');
-        const name = $(this).data('name');
-        const email = $(this).data('email');
-
-        $('#selectedUserId').val(id);
-        $('#selectedUserName').text(name);
-        $('#selectedUserEmail').text(email);
+    $(document).on('click', '.user-result', function(e) {
+        e.preventDefault();
         
+        if ($(this).hasClass('disabled')) {
+            return;
+        }
+        
+        const user = JSON.parse($(this).attr('data-user'));
+        
+        $('#selectedUserId').val(user.id);
+        $('#selectedUserName').text(user.name);
+        $('#selectedUserEmail').text(user.email);
+        $('#selectedUserPhone').text(user.phone ? 'Tel: ' + user.phone : '');
+        $('#selectedUserMemberSince').text('Üyelik Tarihi: ' + user.member_since);
+        
+        // Mevcut kitapları göster
+        let currentBooksHtml = '';
+        if (user.current_books && user.current_books.length > 0) {
+            currentBooksHtml = '<div class="mt-2"><strong>Mevcut Kitapları:</strong><ul class="list-unstyled mb-0">';
+            user.current_books.forEach(function(book) {
+                let bookClass = book.is_overdue ? 'text-danger' : 'text-muted';
+                currentBooksHtml += `
+                    <li class="${bookClass}">
+                        <small>
+                            ${book.title}<br>
+                            <span class="text-muted">Teslim: ${book.due_date}</span>
+                            ${book.is_overdue ? ' <span class="badge bg-danger">Gecikmiş</span>' : ''}
+                        </small>
+                    </li>`;
+            });
+            currentBooksHtml += '</ul></div>';
+        }
+        $('#selectedUserCurrentBooks').html(currentBooksHtml);
+        
+        $('#selectedUserDetails').show();
         $('#userSearch').val('');
-        $('#clearUserSearch').addClass('d-none');
         $('#userSearchResults').hide();
-        $('#selectedUser').show();
         
         checkFormValidity();
     });
 
     // Kullanıcı seçimini temizle
-    $('#clearSelectedUser').click(function() {
+    $('#clearUserSelection').click(function() {
         $('#selectedUserId').val('');
-        $('#selectedUser').hide();
-        $('#userSearch').val('');
-        $('#clearUserSearch').addClass('d-none');
+        $('#selectedUserDetails').hide();
+        $('#userSearch').val('').focus();
         checkFormValidity();
     });
 
-    // Kullanıcı aramasını temizle
-    $('#clearUserSearch').click(function() {
-        $('#userSearch').val('');
-        $('#userSearchResults').hide();
-        $(this).addClass('d-none');
-    });
-
-    // ISBN ile kitap arama
-    $('#searchBookBtn').click(function() {
-        const isbn = $('#isbnSearch').val().trim();
-        if (!isbn) {
-            showBookSearchError('Lütfen bir ISBN numarası girin.');
+    // Kitap arama
+    let bookSearchTimeout;
+    
+    $('#bookSearch').on('keyup', function() {
+        clearTimeout(bookSearchTimeout);
+        const searchTerm = $(this).val().trim();
+        
+        if (searchTerm.length < 2) {
+            $('#bookSearchResults').hide();
             return;
         }
-
-        // Arama başlamadan önce loading göster
-        $('#searchBookBtn i').removeClass('fa-search').addClass('fa-spinner fa-spin');
-        hideBookSearchError();
-
-        $.get(`/staff/books/search/${isbn}`)
-            .done(function(response) {
-                if (response.book) {
-                    const book = response.book;
-                    $('#selectedBookId').val(book.id);
-                    $('#selectedBookTitle').text(book.title);
-                    $('#selectedBookIsbn').text(book.isbn);
-                    $('#selectedBookAuthor').text(response.authors);
-                    $('#selectedBookPublisher').text(book.publisher ? book.publisher.name : '-');
-                    $('#selectedBookCategory').text(book.category ? book.category.name : '-');
-                    $('#selectedBookPages').text(book.pages || '-');
-                    $('#selectedBookYear').text(book.publication_year || '-');
-                    $('#selectedBookAvailableStock').text(response.available_copies);
-                    
-                    $('#selectedBook').show();
-                    hideBookSearchError();
-                    checkFormValidity();
-                }
-            })
-            .fail(function(jqXHR) {
-                let errorMessage = 'Kitap arama sırasında bir hata oluştu.';
-                
-                if (jqXHR.status === 404) {
-                    if (jqXHR.responseJSON && jqXHR.responseJSON.error) {
-                        errorMessage = jqXHR.responseJSON.error;
-                    } else {
-                        errorMessage = 'Kitap bulunamadı.';
+        
+        bookSearchTimeout = setTimeout(function() {
+            $.get("{{ route('staff.borrowings.search.book') }}", { search: searchTerm })
+                .done(function(books) {
+                    if (!books || books.length === 0) {
+                        $('#bookSearchResults').html('<div class="p-3 text-center">Kitap bulunamadı</div>').show();
+                        return;
                     }
-                } else if (jqXHR.responseJSON && jqXHR.responseJSON.error) {
-                    errorMessage = jqXHR.responseJSON.error;
-                }
-                
-                showBookSearchError(errorMessage);
-                console.error('Arama hatası:', jqXHR.responseJSON || jqXHR.statusText);
-            })
-            .always(function() {
-                // Loading'i kaldır
-                $('#searchBookBtn i').removeClass('fa-spinner fa-spin').addClass('fa-search');
+                    
+                    let html = '<div class="list-group">';
+                    books.forEach(function(book) {
+                        html += `
+                            <a href="#" class="list-group-item list-group-item-action book-result" 
+                               data-book='${JSON.stringify(book)}'>
+                                <div class="d-flex justify-content-between align-items-center">
+                                    <div>
+                                        <h6 class="mb-1">${book.title}</h6>
+                                        <small class="text-muted">
+                                            ISBN: ${book.isbn || 'Belirtilmemiş'}<br>
+                                            Yazar: ${book.author}<br>
+                                            Yayınevi: ${book.publisher}
+                                        </small>
+                                    </div>
+                                    <small class="text-success">${book.available_count} kopya mevcut</small>
+                                </div>
+                            </a>`;
+                    });
+                    html += '</div>';
+                    
+                    $('#bookSearchResults').html(html).show();
+                })
+                .fail(function(error) {
+                    console.error('Kitap arama hatası:', error);
+                    $('#bookSearchResults').html('<div class="p-3 text-center text-danger">Arama sırasında bir hata oluştu</div>').show();
+                });
+        }, 500);
+    });
+
+    // Kitap seçimi
+    $(document).on('click', '.book-result', function(e) {
+        e.preventDefault();
+        const book = JSON.parse($(this).attr('data-book'));
+        
+        $('#selectedBookId').val(book.id);
+        $('#selectedBookTitle').text(book.title);
+        $('#selectedBookAuthor').text('Yazar: ' + book.author);
+        $('#selectedBookPublisher').text('Yayınevi: ' + book.publisher);
+        $('#selectedBookIsbn').text('ISBN: ' + (book.isbn || 'Belirtilmemiş'));
+
+        // Stok seçeneklerini güncelle
+        const $stockSelect = $('#stockSelect');
+        $stockSelect.empty().append('<option value="">Kopya seçiniz...</option>');
+        
+        if (book.stocks && book.stocks.length > 0) {
+            book.stocks.forEach(function(stock) {
+                $stockSelect.append(`
+                    <option value="${stock.id}">
+                        ${stock.barcode ? 'Barkod: ' + stock.barcode + ' | ' : ''}
+                        Durum: ${stock.condition || 'Belirtilmemiş'} | 
+                        Raf: ${stock.shelf}
+                    </option>
+                `);
             });
+        }
+
+        $('#selectedBookDetails').show();
+        $('#bookSearchResults').hide();
+        $('#bookSearch').val('');
+        checkFormValidity();
     });
 
     // Kitap seçimini temizle
-    $('#clearSelectedBook').click(function() {
+    $('#clearBookSelection').click(function() {
         $('#selectedBookId').val('');
-        $('#selectedBook').hide();
-        $('#isbnSearch').val('');
-        $('#clearBookSearch').addClass('d-none');
-        hideBookSearchError();
+        $('#stockSelect').val('');
+        $('#selectedBookDetails').hide();
+        $('#bookSearch').val('').focus();
         checkFormValidity();
     });
 
-    // Kitap aramasını temizle
-    $('#clearBookSearch').click(function() {
-        $('#isbnSearch').val('');
-        $('#clearBookSearch').addClass('d-none');
-        hideBookSearchError();
-        checkFormValidity();
-    });
-
-    // Enter tuşu ile arama yapma
-    $('#isbnSearch').on('keypress', function(e) {
-        if (e.which === 13) { // Enter tuşu
-            e.preventDefault();
-            $('#searchBookBtn').click();
+    // Dışarı tıklandığında sonuçları gizle
+    $(document).on('click', function(e) {
+        if (!$(e.target).closest('.user-search-container').length) {
+            $('#userSearchResults').hide();
+        }
+        if (!$(e.target).closest('.book-search-container').length) {
+            $('#bookSearchResults').hide();
         }
     });
 
-    // Kitap arama hata mesajını göster
-    function showBookSearchError(message) {
-        $('#bookSearchFeedback .alert').text(message);
-        $('#bookSearchFeedback').show();
-        $('#selectedBook').hide();
-        $('#selectedBookId').val('');
-    }
-
-    // Kitap arama hata mesajını gizle
-    function hideBookSearchError() {
-        $('#bookSearchFeedback').hide();
-    }
-
-    // Form geçerliliğini kontrol et
+    // Form doğrulama
     function checkFormValidity() {
-        const hasUser = $('#selectedUserId').val() !== '';
-        const hasBook = $('#selectedBookId').val() !== '';
-        const hasBorrowDate = $('#borrow_date').val() !== '';
-        const hasDueDate = $('#due_date').val() !== '';
+        const isUserSelected = $('#selectedUserId').val() !== '';
+        const isBookSelected = $('#selectedBookId').val() !== '';
+        const isStockSelected = $('#stockSelect').val() !== '';
         
-        $('#submitBtn').prop('disabled', !(hasUser && hasBook && hasBorrowDate && hasDueDate));
+        $('#submitBtn').prop('disabled', !(isUserSelected && isBookSelected && isStockSelected));
     }
 
-    // Tarih değişikliklerini izle
-    $('#borrow_date, #due_date').on('change', checkFormValidity);
+    // Stok seçimi değiştiğinde form doğrulaması
+    $('#stockSelect').on('change', checkFormValidity);
+
+    // Form gönderimi
+    $('#borrowingForm').on('submit', function(e) {
+        e.preventDefault();
+        
+        if (!confirm('Ödünç verme işlemini onaylıyor musunuz?')) {
+            return;
+        }
+
+        const $form = $(this);
+        const $submitBtn = $('#submitBtn');
+        const originalHtml = $submitBtn.html();
+        
+        $submitBtn.prop('disabled', true).html('<div class="spinner-border spinner-border-sm" role="status"><span class="visually-hidden">Kaydediliyor...</span></div>');
+
+        $.ajax({
+            url: $form.attr('action'),
+            method: 'POST',
+            data: $form.serialize(),
+            success: function(response) {
+                if (response.success) {
+                    window.location.href = response.redirect || "{{ route('staff.borrowings.index') }}";
+                } else {
+                    alert(response.message || 'Bir hata oluştu');
+                    $submitBtn.prop('disabled', false).html(originalHtml);
+                }
+            },
+            error: function(xhr) {
+                const response = xhr.responseJSON;
+                alert(response?.message || 'İşlem sırasında bir hata oluştu');
+                $submitBtn.prop('disabled', false).html(originalHtml);
+            }
+        });
+    });
 });
-</script>
-@endsection 
+</script> 
