@@ -1,165 +1,116 @@
 @extends('layouts.staff')
 
-@section('title', 'Ceza Yönetimi')
+@section('title', 'Ceza İşlemleri')
 
 @section('content')
 <div class="container-fluid">
     <div class="d-sm-flex align-items-center justify-content-between mb-4">
-        <h1 class="h3 mb-0 text-gray-800">Ceza Yönetimi</h1>
+        <h1 class="h3 mb-0 text-gray-800">Ceza İşlemleri</h1>
     </div>
+
+    @if(session('success'))
+        <div class="alert alert-success">
+            {{ session('success') }}
+        </div>
+    @endif
+
+    @if(session('error'))
+        <div class="alert alert-danger">
+            {{ session('error') }}
+        </div>
+    @endif
 
     <div class="card shadow mb-4">
         <div class="card-header py-3">
             <h6 class="m-0 font-weight-bold text-primary">Ceza Listesi</h6>
         </div>
         <div class="card-body">
-            @if(session('success'))
-                <div class="alert alert-success">
-                    {{ session('success') }}
-                </div>
-            @endif
-
             <div class="table-responsive">
-                <table class="table table-bordered" id="dataTable">
+                <table class="table table-bordered" id="finesTable">
                     <thead>
                         <tr>
-                            <th>Kullanıcı</th>
+                            <th>ID</th>
+                            <th>Üye</th>
                             <th>Kitap</th>
-                            <th>Gecikme</th>
+                            <th>Ceza Türü</th>
                             <th>Tutar</th>
-                            <th>Ödeme Durumu</th>
-                            <th>Ödeme Yöntemi</th>
-                            <th>Referans No</th>
+                            <th>Durum</th>
+                            <th>Tarih</th>
                             <th>İşlemler</th>
                         </tr>
                     </thead>
                     <tbody>
                         @foreach($fines as $fine)
                             <tr>
+                                <td>{{ $fine->id }}</td>
                                 <td>
-                                    {{ $fine->borrowing->user->name }}
-                                    <small class="d-block text-muted">{{ $fine->borrowing->user->email }}</small>
+                                    {{ $fine->user->name }}
+                                    <small class="d-block text-muted">{{ $fine->user->email }}</small>
                                 </td>
                                 <td>
-                                    {{ $fine->borrowing->book->title }}
-                                    <small class="d-block text-muted">
-                                        Teslim: {{ $fine->borrowing->returned_at->format('d.m.Y') }}
-                                    </small>
+                                    {{ $fine->book->title }}
+                                    <small class="d-block text-muted">ISBN: {{ $fine->book->isbn }}</small>
                                 </td>
-                                <td>{{ $fine->borrowing->returned_at->diffInDays($fine->borrowing->due_date) }} gün</td>
-                                <td>{{ number_format($fine->amount, 2) }} TL</td>
                                 <td>
-                                    @if($fine->payment_status === 'paid')
-                                        <span class="badge badge-success">Ödendi</span>
-                                        <small class="d-block">
-                                            {{ $fine->paid_at->format('d.m.Y H:i') }}
-                                        </small>
-                                    @elseif($fine->payment_status === 'pending')
-                                        <span class="badge badge-warning">Beklemede</span>
+                                    @if($fine->type == 'late')
+                                        <span class="badge bg-warning">Gecikme Cezası</span>
+                                    @elseif($fine->type == 'damage')
+                                        <span class="badge bg-danger">Hasar Cezası</span>
+                                        @if($fine->damage_level)
+                                            <small class="d-block mt-1">
+                                                Hasar Seviyesi: 
+                                                @if($fine->damage_level == 'minor')
+                                                    <span class="text-warning">Hafif</span>
+                                                @elseif($fine->damage_level == 'moderate')
+                                                    <span class="text-orange">Orta</span>
+                                                @elseif($fine->damage_level == 'severe')
+                                                    <span class="text-danger">Ağır</span>
+                                                @endif
+                                            </small>
+                                        @endif
+                                    @elseif($fine->type == 'lost')
+                                        <span class="badge bg-dark">Kayıp Cezası</span>
+                                    @endif
+                                    @if($fine->description)
+                                        <small class="d-block">{{ $fine->description }}</small>
+                                    @endif
+                                </td>
+                                <td>{{ number_format($fine->amount, 2) }} ₺</td>
+                                <td>
+                                    @if($fine->status == 'pending')
+                                        <span class="badge bg-warning">Beklemede</span>
+                                    @elseif($fine->status == 'paid')
+                                        <span class="badge bg-success">Ödendi</span>
+                                        @if($fine->paid_at)
+                                            <small class="d-block">{{ $fine->paid_at->format('d.m.Y') }}</small>
+                                        @endif
                                     @else
-                                        <span class="badge badge-secondary">İptal Edildi</span>
+                                        <span class="badge bg-secondary">İptal Edildi</span>
                                     @endif
                                 </td>
+                                <td>{{ $fine->created_at->format('d.m.Y H:i') }}</td>
                                 <td>
-                                    @if($fine->payment_method === 'cash')
-                                        Nakit
-                                    @elseif($fine->payment_method === 'bank_transfer')
-                                        Banka Havalesi
-                                    @else
-                                        -
+                                    @if($fine->status == 'pending')
+                                        <form action="{{ route('staff.fines.mark-as-paid', $fine) }}" method="POST" class="d-inline">
+                                            @csrf
+                                            @method('PUT')
+                                            <button type="submit" class="btn btn-success btn-sm" onclick="return confirm('Cezayı ödenmiş olarak işaretlemek istediğinize emin misiniz?')">
+                                                <i class="fas fa-check"></i> Ödendi
+                                            </button>
+                                        </form>
+                                        
+                                        <form action="{{ route('staff.fines.forgive', $fine) }}" method="POST" class="d-inline">
+                                            @csrf
+                                            @method('PUT')
+                                            <button type="submit" class="btn btn-warning btn-sm" onclick="return confirm('Cezayı affetmek istediğinize emin misiniz?')">
+                                                <i class="fas fa-hand-holding-heart"></i> Affet
+                                            </button>
+                                        </form>
                                     @endif
-                                </td>
-                                <td>{{ $fine->payment_reference ?: '-' }}</td>
-                                <td>
-                                    @if($fine->payment_status === 'pending')
-                                        <button type="button" 
-                                                class="btn btn-primary btn-sm" 
-                                                data-toggle="modal" 
-                                                data-target="#approveModal{{ $fine->id }}">
-                                            <i class="fas fa-check"></i> Onayla
-                                        </button>
-                                        <button type="button" 
-                                                class="btn btn-danger btn-sm"
-                                                data-toggle="modal" 
-                                                data-target="#cancelModal{{ $fine->id }}">
-                                            <i class="fas fa-times"></i> İptal
-                                        </button>
-
-                                        <!-- Onaylama Modalı -->
-                                        <div class="modal fade" id="approveModal{{ $fine->id }}" tabindex="-1">
-                                            <div class="modal-dialog">
-                                                <div class="modal-content">
-                                                    <form action="{{ route('staff.fines.approve', $fine->id) }}" method="POST">
-                                                        @csrf
-                                                        <div class="modal-header">
-                                                            <h5 class="modal-title">Ceza Ödemesi Onaylama</h5>
-                                                            <button type="button" class="close" data-dismiss="modal">
-                                                                <span>&times;</span>
-                                                            </button>
-                                                        </div>
-                                                        <div class="modal-body">
-                                                            <div class="form-group">
-                                                                <label for="payment_method">Ödeme Yöntemi</label>
-                                                                <select class="form-control" name="payment_method" required>
-                                                                    <option value="">Seçiniz</option>
-                                                                    <option value="cash">Nakit</option>
-                                                                    <option value="bank_transfer">Banka Havalesi</option>
-                                                                </select>
-                                                            </div>
-                                                            <div class="form-group">
-                                                                <label for="payment_reference">Ödeme Referansı</label>
-                                                                <input type="text" 
-                                                                       class="form-control" 
-                                                                       name="payment_reference"
-                                                                       placeholder="Dekont no veya makbuz no"
-                                                                       required>
-                                                            </div>
-                                                            <div class="form-group">
-                                                                <label for="admin_notes">Notlar</label>
-                                                                <textarea class="form-control" 
-                                                                          name="admin_notes" 
-                                                                          rows="3"></textarea>
-                                                            </div>
-                                                        </div>
-                                                        <div class="modal-footer">
-                                                            <button type="button" class="btn btn-secondary" data-dismiss="modal">İptal</button>
-                                                            <button type="submit" class="btn btn-primary">Onayla</button>
-                                                        </div>
-                                                    </form>
-                                                </div>
-                                            </div>
-                                        </div>
-
-                                        <!-- İptal Modalı -->
-                                        <div class="modal fade" id="cancelModal{{ $fine->id }}" tabindex="-1">
-                                            <div class="modal-dialog">
-                                                <div class="modal-content">
-                                                    <form action="{{ route('staff.fines.cancel', $fine->id) }}" method="POST">
-                                                        @csrf
-                                                        <div class="modal-header">
-                                                            <h5 class="modal-title">Ceza İptali</h5>
-                                                            <button type="button" class="close" data-dismiss="modal">
-                                                                <span>&times;</span>
-                                                            </button>
-                                                        </div>
-                                                        <div class="modal-body">
-                                                            <div class="form-group">
-                                                                <label for="admin_notes">İptal Nedeni</label>
-                                                                <textarea class="form-control" 
-                                                                          name="admin_notes" 
-                                                                          rows="3"
-                                                                          required></textarea>
-                                                            </div>
-                                                        </div>
-                                                        <div class="modal-footer">
-                                                            <button type="button" class="btn btn-secondary" data-dismiss="modal">Vazgeç</button>
-                                                            <button type="submit" class="btn btn-danger">İptal Et</button>
-                                                        </div>
-                                                    </form>
-                                                </div>
-                                            </div>
-                                        </div>
-                                    @endif
+                                    
+                                    <a href="{{ route('staff.fines.show', $fine) }}" class="btn btn-info btn-sm">
+                                        <i class="fas fa-eye"></i> Detay
+                                    </a>
                                 </td>
                             </tr>
                         @endforeach
@@ -171,14 +122,16 @@
 </div>
 @endsection
 
-@section('scripts')
+@push('scripts')
 <script>
 $(document).ready(function() {
-    $('#dataTable').DataTable({
-        language: {
-            url: '//cdn.datatables.net/plug-ins/1.10.24/i18n/Turkish.json'
-        }
+    $('#finesTable').DataTable({
+        "language": {
+            "url": "//cdn.datatables.net/plug-ins/1.10.24/i18n/Turkish.json"
+        },
+        "order": [[ 0, "desc" ]],
+        "pageLength": 25
     });
 });
 </script>
-@endsection 
+@endpush 
